Restructuring of api integrations
