+++ apiintegrations/jamfpro/interface.go
@@ -10,9 +10,9 @@ import (
 
 // JamfAPIHandler implements the APIHandler interface for the Jamf Pro API.
 type JamfAPIHandler struct {
-	OverrideBaseDomain string        // OverrideBaseDomain is used to override the base domain for URL construction.
-	InstanceName       string        // InstanceName is the name of the Jamf instance.
-	Logger             logger.Logger // Logger is the structured logger used for logging.
+	BaseDomain   string        // OverrideBaseDomain is used to override the base domain for URL construction.
+	InstanceName string        // InstanceName is the name of the Jamf instance.
+	Logger       logger.Logger // Logger is the structured logger used for logging.
 }
 
 type TokenResponse struct {
